backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pytest
